% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSConvert.R
\name{qs.convert}
\alias{qs.convert}
\title{Quantitative Stratigraphy Data Conversion function}
\usage{
qs.convert(input, output, file1, file2, file3)
}
\arguments{
\item{input}{Enter the format of your data as follow: "STANDARD","SINOCOR","CONOP","RASC"}

\item{output}{Enter the format the data you want to get as follow: "STANDARD","SINOCOR","CONOP","RASC"}

\item{file1, file2, file3}{Enter the path to the input data file, file1=.xlsx/.dic/.sct, file2=.dep/.evt,file3=.dat}
}
\description{
This function allows you to convert your data between data formats required by several quantitative stratigraphy software.
The function needs "tidyverse", "xlsx" and "stringr" package as front.
}
\examples{
date2day_fun("STANDARD", "RASC", "C:/Users/Unknown/example.xlsx")

}
\keyword{Quantitative}
\keyword{Stratigraphy}
